+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh3cs9vme9jh1b49x4bwptff"
+        "conv_01kh3kstxdcy1r3afzshwpd653"
     )
     r"""Unique conversation identifier with `conv_` prefix."""