+++ packages/orq-rc/RELEASES.md
@@ -5439,3 +5439,13 @@ Based on:
 - [python v4.3.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.43 - packages/orq-rc
+
+## 2026-02-10 11:13:12
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.3 (2.814.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.45] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.45] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.45 - packages/orq-rc